sokker/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "sokker-ntdb"
version = "0.1.0"
requires-python = ">=3.11"
classifiers = [
    "Private :: Do not upload",
]
dependencies = [
    # core framework
    "Django == 4.2.*",
    "Pillow",
    "Psycopg2",
    "whitenoise",
    "gunicorn",
    # extensions
    "django-debug-toolbar",
    "django-environ",
    "django-sms",
    "wagtail-review",
    # third-party connectors
    "sentry-sdk[django]",
    "django-compressor",
    "django-sass-processor",
    "django-import-export",
    "django-admin-list-filter-dropdown",
    "django-grappelli",
    "django_tables2",
    "django-filter",
    "django-bootstrap-v5",
    "django-admin-numeric-filter",
    "requests",
    "matplotlib",
    "numpy",
    "scikit-learn",
    "pandas",
]

[project.optional-dependencies]
dev = [
    # developer tooling
    "pip-audit",
    "flake8",
    "flake8-pyproject",
    "black",
    "isort",
    # documentation build
    "sphinx",
    "myst-parser",
    "sphinxcontrib-confluencebuilder",
    # testing
    "pytest",
    "pytest-django",
    "pytest-env",
    "pytest-freezer",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sokker"]

[tool.pytest.ini_options]
env = [
    "DJANGO_SETTINGS_MODULE=sokker.settings.tests",
    "SECRET_KEY=6UbOJ4S2Al(YrJ",
]
filterwarnings = [
    "once::DeprecationWarning",
    "once::PendingDeprecationWarning",
]
addopts = "-svv --showlocals --reuse-db --fail-on-template-vars"
testpaths = "sokker"
norecursedirs = [
    ".tox",
    "__pycache__",
    "migrations",
    "templates",
    "jinja2",
]
xfail_strict = true

[tool.isort]
line_length = 99
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
known_django = ["django","celery","constance","rest_framework"]
known_wagtail = ["wagtail"]
sections = ["FUTURE","STDLIB","DJANGO","WAGTAIL","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]

[tool.flake8]
max-line-length = 99
exclude = [".git",".tox","__pycache__","migrations","templates","jinja2"]
show-source = true

[tool.coverage.run]
relative_files = true
